[18] Refactor Kubernetes generators to single package
