+++ src/main/resources/META-INF/services/au.com.agiledigital.toolform.plugin.ToolFormGenerateCommandPlugin
@@ -1,3 +1,3 @@
-au.com.agiledigital.toolform.command.generate.minikube.GenerateMinikubeCommand
+au.com.agiledigital.toolform.command.generate.kubernetes.minikube.GenerateMinikubeCommand
 au.com.agiledigital.toolform.command.generate.docker.GenerateDockerComposeV3Command
-au.com.agiledigital.toolform.command.generate.minishift.GenerateMinishiftCommand
+au.com.agiledigital.toolform.command.generate.kubernetes.minishift.GenerateMinishiftCommand
