+++ src/main/scala/au/com/agiledigital/toolform/command/generate/kubernetes/DeploymentWriter.scala
@@ -1,4 +1,4 @@
-package au.com.agiledigital.toolform.command.generate.minikube
+package au.com.agiledigital.toolform.command.generate.kubernetes
 
 import au.com.agiledigital.toolform.model.ToolFormService
 import cats.implicits._
