+++ src/main/scala/au/com/agiledigital/toolform/command/generate/kubernetes/minishift/GenerateMinishiftCommand.scala
@@ -1,12 +1,13 @@
-package au.com.agiledigital.toolform.command.generate.minikube
+package au.com.agiledigital.toolform.command.generate.kubernetes.minishift
 
 import java.io.{BufferedWriter, File, FileWriter}
 import java.nio.file.Path
 
 import au.com.agiledigital.toolform.app.ToolFormError
-import au.com.agiledigital.toolform.command.generate.minikube.DeploymentWriter.writeDeployment
-import au.com.agiledigital.toolform.command.generate.minikube.GenerateMinikubeCommand.runGenerateMinikube
-import au.com.agiledigital.toolform.command.generate.minikube.ServiceWriter.writeService
+import au.com.agiledigital.toolform.command.generate.kubernetes.DeploymentWriter.writeDeployment
+import au.com.agiledigital.toolform.command.generate.kubernetes.ServiceWriter.writeService
+import au.com.agiledigital.toolform.command.generate.kubernetes.minishift.GenerateMinishiftCommand.runGenerateMinishift
+import au.com.agiledigital.toolform.command.generate.kubernetes.minishift.RouterWriter.writeRouter
 import au.com.agiledigital.toolform.command.generate.{WriterContext, YamlWriter}
 import au.com.agiledigital.toolform.model._
 import au.com.agiledigital.toolform.plugin.ToolFormGenerateCommandPlugin
@@ -23,13 +24,13 @@ import com.monovore.decline.Opts
   *
   * @see https://kubernetes.io/docs/api-reference/v1.8/
   */
-class GenerateMinikubeCommand extends ToolFormGenerateCommandPlugin {
+class GenerateMinishiftCommand extends ToolFormGenerateCommandPlugin {
 
   /**
-    * The primary class for generating Kubernetes (Minikube) config files.
+    * The primary class for generating Kubernetes (Minishift) config files.
     */
   def command: Opts[Either[ToolFormError, String]] =
-    Opts.subcommand("minikube", "generates config files for Kubernetes (Minikube) container orchestration") {
+    Opts.subcommand("minishift", "generates config files for Kubernetes (Minishift) container orchestration") {
       (Opts.option[Path]("in-file", short = "i", metavar = "file", help = "the path to the project config file") |@|
         Opts.option[Path]("out-file", short = "o", metavar = "file", help = "the path to output the generated file(s)"))
         .map(execute)
@@ -47,23 +48,23 @@ class GenerateMinikubeCommand extends ToolFormGenerateCommandPlugin {
     } else {
       for {
         project <- ProjectReader.readProject(inputFile)
-        status  <- runGenerateMinikube(inputFile.getAbsolutePath, outputFile, project)
+        status  <- runGenerateMinishift(inputFile.getAbsolutePath, outputFile, project)
       } yield status
     }
   }
 }
 
-object GenerateMinikubeCommand extends YamlWriter {
+object GenerateMinishiftCommand extends YamlWriter {
 
   /**
-    * The main entry point into the Kubernetes (Minikube) file generation.
+    * The main entry point into the Kubernetes (Minishift) file generation.
     *
     * @param sourceFilePath project config input file path
     * @param project        the abstract project definition parsed by ToolFormApp.
     * @return on success it returns a status message to print to the screen, otherwise it will return an
     *         error object describing what went wrong.
     */
-  def runGenerateMinikube(sourceFilePath: String, outFile: File, project: Project): Either[ToolFormError, String] = {
+  def runGenerateMinishift(sourceFilePath: String, outFile: File, project: Project): Either[ToolFormError, String] = {
     val writer = new BufferedWriter(new FileWriter(outFile, false))
     try {
       val writeFile = for {
@@ -74,6 +75,7 @@ object GenerateMinikubeCommand extends YamlWriter {
         _ <- project.resources.values.filter(shouldWriteService).toList.traverse_(writeService)
         _ <- project.components.values.toList.traverse_((component) => writeDeployment(project.id, component))
         _ <- project.resources.values.toList.traverse_((resource) => writeDeployment(project.id, resource))
+        _ <- project.topology.endpoints.toList.traverse_ { case (endpointId, endpoint) => writeRouter(endpointId, endpoint) }
       } yield ()
 
       val context = WriterContext(writer)
